requirements.txt
fastapi==0.95.2
pydantic==1.10.15
sqlalchemy==2.0.38
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
uvicorn==0.22.0
email-validator==2.0.0
httpx
bcrypt==3.2.2
